2.3.5 - MEL change for Maya 2009 - compatibility fixes for linux - output of plugin shapes in RIB, so they can be used as placeholders
2.3.4
- RIReadArchive Mel command updated, the -inline(i) flag inlines the Read Archive into the RIB
- liquidCleanup command script modified so confirm dialog can be supressed (to use it in scripts)
- wiki link changed in globals
- "Render All Curves" global option fixed, global switch for curve rendering
- the Maya Subdiv render module is fixed, maya subdivs are now ouput properly, with the following limitations:
  - only the base mesh (level 0) is output (no hierarchy edits)
  - level 0 edge/vertex creases are always hard (crease value 7), all other levels are ignored
 - maya subdivs can be tagged like polygons through the liquid menu which gives access to the boundary interpolation options

+++ trunk/include/liqRibMayaSubdivisionData.h
@@ -90,7 +90,7 @@ private: // Data
 
   void checkExtraTags( MObject &mesh );
   void addExtraTags( MObject &mesh, SBD_EXTRA_TAG extraTag );
-  void addExtraTags( MObject &mesh, float extraTagValue, SBD_EXTRA_TAG extraTag );
+  void addExtraTags( MObject &mesh, int extraTagValue, SBD_EXTRA_TAG extraTag );
 };
 
 #endif
